Fail Log
Fail Log for Module Two
July 17 - 23: Module 2 Finding Data
Previous Mistakes
A few mistakes that I think are important to note was that I completed the blog post wrong. I misunderstand the part of Dillinger and did the post in the software rather than on Word Press. This created a lot of issues as the post was not able to be seen when I submitted the assignment.

Another mistake that I noticed was I tried to format with Dillinger into my fail logs, however it didn’t work. This week, I didn’t make the same mistake and made sure it set it up correctly.

Another comment that was made was how I didn’t expand my relation to the Readings, so this week, I’m explaining in a lot more detail on why I can relate to them and how.

Readings
For this week’s annotations, we had a few choices on which articles to do our annotations on. I choose to do mine on Seams and Edges: Dreams of Aggregation, Access & Discovery in a Broken World. To summarize, the reading goes over the process of technolog